3.2.11

+++ src/components/Task/Task.jsx
@@ -1,60 +1,47 @@
 /* eslint-disable react/destructuring-assignment */
 import { formatDistanceToNow } from 'date-fns'
-import { Component, createRef } from 'react'
+import { useState, useRef, useEffect } from 'react'
 
-class Task extends Component {
-  state = {
-    inputValueEdit: this.props.item,
-    editingToggle: false,
-    timer: this.props.userTimerSeconds,
-    timerStart: false,
-  }
+function Task({
+  item,
+  userTimerSeconds,
+  dataCreated,
+  taskComplete,
+  handleDeleteTask,
+  id,
+  list,
+  handleCompletedTasks,
+  completedFlag,
+  activeFlag,
+}) {
+  const [inputValueEdit, setInputValueEdit] = useState(item)
+  const [changeInputValueEdit, setChangeInputValueEdit] = useState(item)
+  const [editingToggle, setEditingToggle] = useState(false)
+  const [timer, setTimer] = useState(userTimerSeconds)
+  const [timerStart, setTimerStart] = useState(false)
+  const inputRef = useRef()
 
-  componentDidMount() {
-    this.timerItem = setInterval(() => this.isTimer(), 1000)
-    this.inputRef = createRef()
-  }
-
-  componentDidUpdate() {
-    this.inputRef.current.focus()
-  }
-
-  componentWillUnmount() {
-    clearInterval(this.timerItem)
-  }
-
-  handleInputKeyDown = (e) => {
-    const { value } = e.target
-    if (e.key === 'Enter' && value.trim().length !== 0) {
-      this.setState({ editingToggle: false, inputValueEdit: value })
+  const isTimer = () => {
+    if (userTimerSeconds === 0) {
+      setTimer((prevTimer) => prevTimer + 1)
+    } else {
+      setTimer((prevTimer) => prevTimer - 1)
     }
   }
 
-  timeCreation = () => {
-    const { dataCreated } = this.props
-    const time = formatDistanceToNow(dataCreated, {
-      includeSeconds: true,
-      addSuffix: true,
-    })
-    return time
-  }
-
-  isTimer = () => {
-    const { timerStart } = this.state
-    const { userTimerSeconds } = this.props
+  useEffect(() => {
+    let timerItem
     if (timerStart) {
-      if (userTimerSeconds === 0) {
-        this.setState((prevState) => ({
-          timer: prevState.timer + 1,
-        }))
-      } else {
-        this.setState((prevState) => ({ timer: prevState.timer - 1 }))
-      }
+      timerItem = setInterval(() => isTimer(), 1000)
     }
-  }
+    return () => clearInterval(timerItem)
+  }, [timerStart])
 
-  formatTimer = () => {
-    const { timer } = this.state
+  useEffect(() => {
+    inputRef.current.focus()
+  }, [editingToggle])
+
+  const formatTimer = () => {
     if (timer >= 0) {
       return `${Math.floor(timer / 60) < 10 ? '0' : ''}${Math.floor(timer / 60)} : ${timer % 60 < 10 ? '0' : ''}${
         timer % 60
@@ -65,40 +52,45 @@ class Task extends Component {
     }${Math.abs(timer % 60)}`
   }
 
-  handleOnClick = (e) => {
-    const { ariaLabel } = e.target
-    if (ariaLabel === 'Play') {
-      this.setState({
-        timerStart: true,
-      })
-    } else if (ariaLabel === 'Pause') {
-      this.setState({
-        timerStart: false,
-      })
+  const handleInputKeyDown = (e) => {
+    if (e.key === 'Enter' && changeInputValueEdit.trim().length !== 0) {
+      setEditingToggle(false)
+      setInputValueEdit(changeInputValueEdit)
+    }
+    if (e.key === 'Escape') {
+      setEditingToggle(false)
+      setChangeInputValueEdit(inputValueEdit)
     }
   }
 
-  toggleEditTask = () => {
-    if (!this.props.taskComplete) this.setState({ editingToggle: true })
+  const timeCreation = () => {
+    const time = formatDistanceToNow(dataCreated, {
+      includeSeconds: true,
+      addSuffix: true,
+    })
+    return time
   }
 
-  deleteTask = () => {
-    const { handleDeleteTask, id, list } = this.props
-    const index = list.findIndex((el) => el[2] === id)
-    const tasksItemDelete = list.filter((_, ind) => ind !== index)
-    handleDeleteTask(tasksItemDelete)
+  const handleOnClick = (e) => {
+    const { ariaLabel } = e.target
+    if (ariaLabel === 'Play') {
+      setTimerStart(true)
+    } else if (ariaLabel === 'Pause') {
+      setTimerStart(false)
+    } else if (ariaLabel === 'Edit') {
+      if (!taskComplete) setEditingToggle(true)
+    } else if (ariaLabel === 'Destroy') {
+      const index = list.findIndex((el) => el[2] === id)
+      const tasksItemDelete = list.filter((_, ind) => ind !== index)
+      handleDeleteTask(tasksItemDelete)
+    } else if (ariaLabel === 'Completed') {
+      const index = list.findIndex((el) => el[2] === id)
+      const newList = list.map((it, ind) => (ind !== index ? it : [...it.slice(0, 4), !it[4]]))
+      handleCompletedTasks(newList)
+    }
   }
 
-  completedTask = () => {
-    const { handleCompletedTasks, id, list } = this.props
-    const index = list.findIndex((el) => el[2] === id)
-    const newList = list.map((item, ind) => (ind !== index ? item : [...item.slice(0, 4), !item[4]]))
-    handleCompletedTasks(newList)
-  }
-
-  liClassName = () => {
-    const { taskComplete, completedFlag, activeFlag } = this.props
-    const { editingToggle } = this.state
+  const liClassName = () => {
     let className = ''
     if (taskComplete) {
       className = 'completed'
@@ -113,46 +105,43 @@ class Task extends Component {
     return className
   }
 
-  render() {
-    const { id, taskComplete } = this.props
-    const { inputValueEdit, timerStart, timer } = this.state
-    return (
-      <li className={this.liClassName()}>
-        <div className="view">
-          <input
-            className="toggle"
-            type="checkbox"
-            onClick={this.completedTask}
-            defaultChecked={!!taskComplete}
-            id={id}
-          />
-          <label htmlFor={id}>
-            <span className="title">{inputValueEdit}</span>
-            <span className={timer >= 0 ? 'description' : 'timer-expired'}>
-              <button
-                className={timerStart ? 'icon icon-pause' : 'icon icon-play'}
-                aria-label={timerStart ? 'Pause' : 'Play'}
-                type="button"
-                onClick={this.handleOnClick}
-              />
-              {this.formatTimer()}
-            </span>
-            <span className="description">created {this.timeCreation()}</span>
-          </label>
-          <button className="icon icon-edit" onClick={this.toggleEditTask} aria-label="Edit" type="button" />
-          <button className="icon icon-destroy" onClick={this.deleteTask} aria-label="Destroy" type="button" />
-        </div>
+  return (
+    <li className={liClassName()}>
+      <div className="view">
         <input
-          type="text"
-          className="edit"
-          onKeyDown={this.handleInputKeyDown}
-          onChange={(e) => this.setState({ inputValueEdit: e.target.value })}
-          value={inputValueEdit}
-          ref={this.inputRef}
+          className="toggle"
+          type="checkbox"
+          onClick={handleOnClick}
+          defaultChecked={!!taskComplete}
+          aria-label="Completed"
+          id={id}
         />
-      </li>
-    )
-  }
+        <label htmlFor={id}>
+          <span className="title">{inputValueEdit}</span>
+          <span className={timer >= 0 ? 'description' : 'timer-expired'}>
+            <button
+              className={timerStart ? 'icon icon-pause' : 'icon icon-play'}
+              aria-label={timerStart ? 'Pause' : 'Play'}
+              type="button"
+              onClick={handleOnClick}
+            />
+            {formatTimer()}
+          </span>
+          <span className="description"> created {timeCreation()}</span>
+        </label>
+        <button className="icon icon-edit" onClick={handleOnClick} aria-label="Edit" type="button" />
+        <button className="icon icon-destroy" onClick={handleOnClick} aria-label="Destroy" type="button" />
+      </div>
+      <input
+        type="text"
+        className="edit"
+        onKeyDown={handleInputKeyDown}
+        onChange={(e) => setChangeInputValueEdit(e.target.value)}
+        value={changeInputValueEdit}
+        ref={inputRef}
+      />
+    </li>
+  )
 }
 
 export default Task
